specimens/pygemini/pyproject.toml
[project]
name = "pygemini"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "gitpython>=3.1.46",
    "google-genai>=1.69.0",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "rich>=14.3.3",
    "tiktoken>=0.12.0",
    "watchfiles>=1.1.1",
]

[project.scripts]
pygemini = "pygemini.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pygemini"]
